Created version 2 of the programs using int8_t and int16_t from the <inttypes.h> library. Also add the int return for main, the void for no args for main, and return code for the program.

+++ The_ANSI_C_Programming_Language/Chapter01/fahrenheit_celcius_table_float2.c
@@ -1,10 +1,12 @@
 #include <stdio.h>
+// import the new int types library
+#include <inttypes.h>
 
 // void here means don't take any arguments
 int main(void)
 {
 	float f, c;
-	int8_t lower, upper, increment;
+	int16_t lower, upper, increment;
 
 	lower = 0;
 	upper = 300;
